clementine_tx_sender

Module citrea

Provides functions to build Bitcoin transactions related to commit-reveal pattern for Citrea rollup.

Modules

data_serialization

reveal_scripts 🔒

This module contains functions to create transactions for the DA layer.

sync

Enums

CitreaTxRequest

Citrea DA payload request.

TransactionKind

Type represents a typed enum for transaction kind Conversion to u16 (to_bytes) should be same as used in citrea repo. citrea/crates/bitcoin-da/src/helpers/mod.rs

Constants

MAX_CHUNK_SIZE 🔒

Maximum raw BatchProof chunk size.

Functions

build_commit_transaction 🔒

Build the commit part of commit-reveal pair Multiple commits can be in the same tx (if chunks are used, each commit needs a different nonce so that the addresses are different)

build_reveal_transaction 🔒

Build the reveal part of commit-reveal pair

calculate_sha256 🔒

sign_blob_with_private_key

Signs a message with a private key Returns (signature, public_key)