clementine_tx_sender::citrea

Module data_serialization

Structs

BatchProofMethodId 🔒

A new batch proof method_id starting to be applied from the l2_block_number (inclusive).

BatchProofMethodIdBody 🔒

Body of the batch proof method id update for light client

SequencerCommitment 🔒

Commitments made to the DA layer from the sequencer. Has merkle root of l2 block hashes from L1 start block to L1 end block (inclusive)

Enums

DataOnDa 🔒

Data written to DA and read from DA must be the borsh serialization of this enum WARNING: This type has to be the exact same as the one used in citrea repo, so that the borsh serializations are the same. citrea/crates/sovereign-sdk/rollup-interface/src/state_machine/da.rs

Constants

SECURITY_COUNCIL_SIGNATURE_SIZE 🔒

Size of a signature in bytes.

SECURITY_COUNCIL_SIGNATURE_THRESHOLD 🔒

Minimum number of verified signatures required to approve a method id upgrade.

Type Aliases

Proof 🔒

The ZK proof generated by the [ZkvmHost::run] method.