clementine_tx_sender

Module cpfp

Child Pays For Parent (CPFP) Support For Transaction Sender

This module implements the Child Pays For Parent (CPFP) strategy for sending Bitcoin transactions with transaction sender.

Child Transaction Details

A child transaction is created to pay for the fees of a parent transaction. They must be submitted together as a package for Bitcoin nodes to accept them.

Fee Payer Transactions/UTXOs

Child transaction needs to spend an UTXO for the fees. But because of the TRUC rules (https://github.com/bitcoin/bips/blob/master/bip-0431.mediawiki#specification), a third transaction can’t be put into the package. So, a so called “fee payer” transaction must be send and confirmed before the CPFP package is send.