clementine_tx_sender

Module test_utils

Test utilities for clementine-tx-sender.

Structs

WithProcessCleanup

Functions

create_regtest_rpc

Creates a Bitcoin regtest node for testing, waits for it to start and returns an RPC client.

create_test_environment

Creates a test environment with a unique database name and a regtest Bitcoin node specific to the test.

get_available_port

Helper to get a dynamically assigned free port.

get_current_test_name

setup_txsender_test_db

Sets up a test database with a unique name based on the current test thread name.