Crate clementine_tx_sender

Clementine Transaction Sender

This crate handles the creation, signing, and broadcasting of Bitcoin transactions, supporting various fee-bumping strategies like CPFP and RBF.

Re-exports

pub use client::TxSenderClient;

Modules

client

Transaction Sender Client

cpfp

Child Pays For Parent (CPFP) Support For Transaction Sender

nonstandard

rbf

task

Macros

log_error_for_tx

Structs

ActivatedWithOutpoint

Activation condition based on an outpoint.

ActivatedWithTxid

Activation condition based on a transaction ID.

MempoolConfig

SpendableUtxo

Represents a spendable UTXO.

TxSender

Manages the process of sending Bitcoin transactions, including handling fee bumping strategies like Replace-By-Fee (RBF) and Child-Pays-For-Parent (CPFP).

Enums

SendTxError

Transaction sending errors.

Constants

DEFAULT_SEQUENCE

Default sequence for transactions.

Traits

SpendableInputInfo

Trait for extracting information from a spendable input. This allows different input types (SpendableUtxo, SpendableTxIn) to be used interchangeably.

TxSenderDatabase

Trait for database operations required by the transaction sender.

TxSenderSigner

Trait for signing transactions in the transaction sender.

TxSenderTxBuilder

Trait for building child transactions in the transaction sender.

Type Aliases

Result